0000997: dbimport throttle

+++ symmetric-io/src/main/java/org/jumpmind/symmetric/io/data/writer/DatabaseWriterSettings.java
@@ -33,7 +33,7 @@ public class DatabaseWriterSettings {
     protected long maxRowsBeforeCommit = 10000;
     
     // Milliseconds to sleep between commits.
-    protected int commitSleepInterval = 5;
+    protected long commitSleepInterval = 5;
 
     protected boolean treatDateTimeFieldsAsVarchar = false;
 
@@ -235,11 +235,11 @@ public class DatabaseWriterSettings {
 
     }
 
-    public int getCommitSleepInterval() {
+    public long getCommitSleepInterval() {
         return commitSleepInterval;
     }
 
-    public void setCommitSleepInterval(int commitSleepInterval) {
+    public void setCommitSleepInterval(long commitSleepInterval) {
         this.commitSleepInterval = commitSleepInterval;
     }
